% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accessions_long}
\alias{accessions_long}
\title{GenBank accessions used in the Fern Tree of Life (FTOL), long format}
\format{
A tibble (data frame) with 15023 rows and
8 columns.
\describe{
\item{species}{Species name; matches names of tips in tree}
\item{locus}{Name of locus (gene or intergenic spacer region)}
\item{accession}{GenBank accession number}
\item{seq_len}{Sequence length (bp), excluding any missing or ambiguous bases}
\item{sci_name}{Scientific name used in FTOL}
\item{ncbi_name}{Scientific name used in the NCBI taxonomic database}
\item{ncbi_taxid}{NCBI taxonomy database unique identifier}
\item{outgroup}{Logical; TRUE for outgroup taxa, FALSE for ingroup
taxa (ferns)}
}
}
\usage{
accessions_long
}
\description{
One row per accession.
}
\details{
For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). https://doi.org/10.1101/2022.03.31.486640
}
\keyword{datasets}
