% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{accessions_wide}
\alias{accessions_wide}
\title{GenBank accessions used in the Fern Tree of Life (FTOL), wide format}
\format{
A tibble (data frame) with 5612 rows and
13 columns.
\describe{
\item{species}{Species name; matches names of tips in tree}
\item{atpA}{GenBank accession number for atpA}
\item{atpB}{GenBank accession number for atpB}
\item{matK}{GenBank accession number for matK}
\item{rbcL}{GenBank accession number for rbcL}
\item{rps4}{GenBank accession number for rps4}
\item{rps4-trnS}{GenBank accession number for rps4-trnS}
\item{trnL-trnF}{GenBank accession number for trnL-trnF}
\item{plastome}{GenBank accession number for plastomes}
\item{join_by}{Method used to join loci}
\item{specimen_voucher}{Specimen voucher}
\item{publication}{Publication}
\item{outgroup}{Logical; TRUE for outgroup taxa, FALSE for ingroup
taxa (ferns)}
}
}
\usage{
accessions_wide
}
\description{
One row per species.
}
\details{
\code{join_by} only provided for Sanger (non-plastome) sequences.
\code{specimen_voucher} and \code{publication} only provided for Sanger
(non-plastome) sequences, and only if these were the same across all loci.
\code{specimen_voucher} has been converted to lower case, alpha-numeric
characters only to account for orthographic variation in GenBank data.

For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life.
https://doi.org/10.3389/fpls.2022.909768
}
\keyword{accessions}
\keyword{data}
