% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_seqs.R
\name{ft_seqs}
\alias{ft_seqs}
\title{Load the DNA sequences used to build the Fern Tree of Life (FTOL)}
\usage{
ft_seqs(
  loci = NULL,
  plastome = FALSE,
  aligned = TRUE,
  del_gaps = ifelse(is.null(loci), FALSE, TRUE),
  drop_og = FALSE
)
}
\arguments{
\item{loci}{Character vector (optional); names of loci to include in the
output. If \code{NULL} (default), all loci will be included. For a list of valid
locus names, see \link{plastome_parts} for plastome loci and \link{sanger_parts} for
Sanger loci.}

\item{plastome}{Logical vector of length 1; If \code{TRUE}, the alignment will
include loci and species with complete plastomes available; otherwise
will include Sanger loci and all species. Default \code{FALSE}.}

\item{aligned}{Logical vector of length 1; If \code{TRUE}, the output will be
aligned sequences (matrix); otherwise will be unaligned (list).
Default \code{TRUE}.}

\item{del_gaps}{Logical vector of length 1; If \code{TRUE}, all gaps
will be deleted from the alignment. Default is \code{TRUE} if \code{loci} is \code{NULL},
\code{FALSE} otherwise.}

\item{drop_og}{Logical vector of length 1; If \code{TRUE}, the outgroup
(non-ferns) will be excluded; otherwise the outgroup is included. Default
\code{FALSE}.}
}
\value{
List or matrix of class "DNAbin"; DNA sequences.
}
\description{
For details about the data, see \link{plastome_alignment} for plastome sequences
and \link{sanger_alignment} for Sanger sequences.
}
\details{
After subsetting loci, columns / rows consisting of only gaps
will be deleted regardless of \code{del_gaps} argument (\code{del_gaps} deletes \emph{all}
gaps, typically resulting in unaligned sequences).

For details on methods used to assemble alignments, see Nitta et al. 2022.
}
\examples{
# Default is aligned Sanger sequences
ft_seqs()
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). https://doi.org/10.1101/2022.03.31.486640
}
