% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_tree.R
\name{ft_tree}
\alias{ft_tree}
\title{Load the Fern Tree of Life (FTOL)}
\usage{
ft_tree(
  sampling = c("backbone", "full"),
  bl = c("ultra", "raw", "clado"),
  drop_og = FALSE
)
}
\arguments{
\item{sampling}{Character vector of length 1; sampling scheme to use when
loading tree. Must choose either \code{"backbone"} (the backbone phylogeny,
including only species with whole plastome data avaiable) or \code{"full"} (the
full phylogeny).}

\item{bl}{Character vector of length 1; how to treat branch lengths. Must
choose from \code{"ultra"} (ultrametric tree, branchlengths in units of time),
\code{"raw"} (raw branchlengths in units of genetic change), or \code{"clado"}
(cladogram, no branchlengths). \code{"ultra"} is only available if \code{sampling} is
\code{"full"}.}

\item{drop_og}{Logical vector of length 1; If \code{TRUE}, the outgroup will be
excluded.}
}
\value{
List of class "phylo"; a phylogenetic tree.
}
\description{
For details on methods used to infer the tree, see Nitta et al. 2022.
}
\examples{
ft_tree("full", "ultra")
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). FIXME ADD DOI
}
