% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_con_dated_tree}
\alias{ftol_con_dated_tree}
\title{Fern Tree of Life (FTOL) dated consensus phylogeny}
\format{
List of class "phylo"
}
\usage{
ftol_con_dated_tree
}
\description{
Majority-rule extended consensus phylogenetic tree of ferns inferred from
seven loci by maximum-likelihood and dated using 60 fossil ferns as
constraints, including maximal sampling of available species on GenBank.
}
\details{
A \link[=backbone_tree]{backbone phylogeny} based on full plastome sequences
including
441
tips was first inferred and used as a constraint tree for the consensus tree.

Includes
5,600
tips and
5,599
internal nodes. Branch lengths are in units of time (Ma). Node labels are
bootstrap support (BS) values (\%).

For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life. https://doi.org/10.3389/fpls.2022.909768
}
\keyword{datasets}
