% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_ferns}
\alias{ftol_ferns}
\title{Fern Tree of Life (FTOL) fern species}
\format{
Character vector
}
\usage{
ftol_ferns
}
\description{
Fern species names included in FTOL in alphabetical order. All species names
are unique. Includes 5563 species.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). https://doi.org/10.1101/2022.03.31.486640
}
\keyword{datasets}
