% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_match_results}
\alias{ftol_match_results}
\title{Fern Tree of Life (FTOL) taxonomic name matching results}
\format{
A tibble (data frame) with 6474 rows and
7 columns.
\describe{
\item{query}{Queried taxonomic name from NCBI}
\item{resolved_name}{Resolved name used in FTOL}
\item{matched_name}{Name matching query in pteridocat}
\item{resolved_status}{Taxonomic status of resolved name}
\item{matched_status}{Taxonomic status of matched name}
\item{match_type}{Type of match assigned by taxontools}
\item{taxid}{NCBI taxonomic ID}
}
}
\usage{
ftol_match_results
}
\description{
Results of taxonomic name matching and resolution for species included in
FTOL. Original names from the NCBI taxonomic database
(https://www.ncbi.nlm.nih.gov/taxonomy) were matched to those in pteridocat
(https://github.com/fernphy/pteridocat), and synonyms resolved to their
accepted names in pteridocat. For compact printing, it is recommended to use
the \code{tibble} library (\code{library(tibble)}).
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life.
https://doi.org/10.3389/fpls.2022.909768

Pteridophyte Phylogeny Group I (2016) A community-derived
classification for extant lycophytes and ferns. Journal of Systematics and
Evolution 54:563–603. \url{https://doi.org/10.1111/jse.12229}
}
\keyword{data}
\keyword{taxonomy}
