% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_ml_tree}
\alias{ftol_ml_tree}
\title{Fern Tree of Life (FTOL) maximum-likelihood phylogeny}
\format{
List of class "phylo"
}
\usage{
ftol_ml_tree
}
\description{
A maximum-likelihood phylogenetic tree of ferns inferred from seven loci,
including maximal sampling of available species on GenBank.
}
\details{
Includes 5581 tips and 5580
internal nodes. Branch lengths are in units of estimated genetic change per
site (not ultrametric). Node labels are bootstrap support (\%).

For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). FIXME ADD DOI
}
\keyword{datasets}
