% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plastome_alignment}
\alias{plastome_alignment}
\title{Aligned plastome DNA sequences used to build Fern Tree of Life (FTOL)}
\format{
Matrix of class "DNAbin"
}
\usage{
plastome_alignment
}
\description{
Concatenation of 79 loci from whole
fern plastomes, including maximal sampling of available species on GenBank.
}
\details{
Includes 442
species and
74,883
bases.

For details on methods used to assemble alignment, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life. https://doi.org/10.3389/fpls.2022.909768
}
\keyword{alignments}
\keyword{data}
