% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sanger_parts}
\alias{sanger_parts}
\title{Start and end positions of loci in Sanger DNA sequence alignment}
\format{
A tibble (data frame) with 7 rows and
3 columns.
\describe{
\item{locus}{Name of locus (gene or intergenic spacer region)}
\item{start}{Start position (column number) of locus in concatenated DNA
alignment}
\item{end}{End position (column number) of locus in concatenated DNA
alignment}
}
}
\usage{
sanger_parts
}
\description{
These data can be used to extract portions of the Sanger DNA sequence
alignment (\link{sanger_alignment}) corresponding to individual loci.
}
\details{
For details on methods used to assemble alignment, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life. https://doi.org/10.3389/fpls.2022.909768
}
\keyword{alignments}
\keyword{data}
