% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{backbone_tree}
\alias{backbone_tree}
\title{Fern Tree of Life (FTOL) backbone phylogeny}
\format{
List of class "phylo"
}
\usage{
backbone_tree
}
\description{
Majority-rule extended consensus phylogenetic tree inferred from whole fern
plastomes by maximum-likelihood, including maximal sampling of available
species on GenBank.
}
\details{
Includes
429
tips and
428
internal nodes. Branch lengths are in units of estimated genetic change per
site (not ultrametric). Node labels are bootstrap support (\%).

For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). FIXME ADD DOI
}
\keyword{datasets}
