% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_data_ver.R
\name{ft_data_ver}
\alias{ft_data_ver}
\title{Current Fern Tree of Life (FTOL) data version}
\usage{
ft_data_ver()
}
\value{
Character vector of length 1; the current FTOL data version loaded
by this package
}
\description{
Current Fern Tree of Life (FTOL) data version
}
\examples{
ft_data_ver()
}
