% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ft_tree.R
\name{ft_tree}
\alias{ft_tree}
\title{Load the Fern Tree of Life (FTOL)}
\usage{
ft_tree(
  branch_len = "ultra",
  backbone = FALSE,
  consensus = FALSE,
  drop_og = FALSE,
  rooted = TRUE
)
}
\arguments{
\item{branch_len}{Character vector of length 1; how to treat branch lengths.
Must choose from \code{"ultra"} (ultrametric tree, branchlengths in units of
time), \code{"raw"} (raw branchlengths in units of genetic change), or \code{"clado"}
(cladogram, no branchlengths).  Default \code{"ultra"}.}

\item{backbone}{Logical vector of length 1; If \code{TRUE}, the backbone phylogeny
(only species with complete plastomes available) will be returned;
otherwise the phylogeny will include all species. Default \code{FALSE}.}

\item{consensus}{Logical vector of length 1; If \code{TRUE}, the majority-rule
extended consensus phylogeny will be returned; otherwise returns the
maximum-likelihood tree. Default \code{FALSE}.}

\item{drop_og}{Logical vector of length 1; If \code{TRUE}, the outgroup
(non-ferns) will be excluded; otherwise the outgroup is included. Default
\code{FALSE}.}

\item{rooted}{Logical vector of length 1; If \code{TRUE}, the phylogeny will be
rooted on bryophytes; otherwise the phylogeny is unrooted. Default \code{TRUE}.}
}
\value{
List of class "phylo"; a phylogenetic tree.
}
\description{
For details on methods used to infer the tree, see Nitta et al. 2022.
}
\details{
Not all combinations are possible. For example, \code{branch_len = "ultra"} is
only available if \code{backbone = FALSE}.
}
\examples{
# Default is the maximum-likelihood tree with branchlengths in units of time
ft_tree()
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). FIXME ADD DOI
}
