% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_con_tree}
\alias{ftol_con_tree}
\title{Fern Tree of Life (FTOL) consensus phylogeny}
\format{
List of class "phylo"
}
\usage{
ftol_con_tree
}
\description{
Majority-rule extended consensus phylogenetic tree of ferns inferred from
seven loci by maximum-likelihood, including maximal sampling of available
species on GenBank.
}
\details{
A \link[=backbone_tree]{backbone phylogeny} based on full plastome sequences
including
428
tips was first inferred and used as a constraint tree for the consensus tree.

Includes
5,581
tips and
5,580
internal nodes. Branch lengths are in units of estimated genetic change per
site (not ultrametric). Node labels are bootstrap support (BS) values (\%).

For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). FIXME ADD DOI
}
\keyword{datasets}
