% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_data_version}
\alias{ftol_data_version}
\title{Fern Tree of Life (FTOL) data version}
\format{
Character vector of length 1.
}
\usage{
ftol_data_version
}
\description{
The version number of the FTOL data included in this R package.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). https://doi.org/10.1101/2022.03.31.486640
}
\keyword{datasets}
