% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_ferns}
\alias{ftol_ferns}
\title{Fern Tree of Life (FTOL) fern species}
\format{
Character vector
}
\usage{
ftol_ferns
}
\description{
Fern species names included in FTOL in alphabetical order. All species names
are unique. Includes 5592 species.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life.
https://doi.org/10.3389/fpls.2022.909768
}
\keyword{data}
\keyword{taxonomy}
