% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_ml_dated_tree}
\alias{ftol_ml_dated_tree}
\title{Fern Tree of Life (FTOL) dated maximum-likelihood phylogeny}
\format{
List of class "phylo"
}
\usage{
ftol_ml_dated_tree
}
\description{
A maximum-likelihood (ML) phylogenetic tree of ferns inferred from seven loci
and dated using 60 fossil ferns as constraints, including maximal sampling of
available species on GenBank.
}
\details{
A \link[=backbone_tree]{backbone phylogeny} based on full plastome
sequences including
428
tips was first inferred and used as a constraint tree for the ML tree.

Includes
5,581
tips and
5,580
internal nodes. Branch lengths are in units of time (Ma). Node labels are
bootstrap support (BS) values (\%) separated by a forward slash. BS value
before the slash is BS calculated from the full (Sanger) sampling; BS value
after the slash is BS calculated from the backbone (plastome) sampling.

For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated Fern Tree of Life (FTOL). FIXME ADD DOI
}
\keyword{datasets}
