% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ftol_ml_tree}
\alias{ftol_ml_tree}
\title{Fern Tree of Life (FTOL) maximum-likelihood phylogeny}
\format{
List of class "phylo"
}
\usage{
ftol_ml_tree
}
\description{
A maximum-likelihood (ML) phylogenetic tree of ferns inferred from seven
loci, including maximal sampling of available species on GenBank.
}
\details{
A \link[=backbone_tree]{backbone phylogeny} based on full plastome sequences
including
503
tips was first inferred and used as a constraint tree for the ML tree.

Includes
5,611
tips and
5,610
internal nodes. Branch lengths are in units of estimated genetic change per
site (not ultrametric). Node labels are bootstrap support (BS) values (\%)
separated by a forward slash. BS value before the slash is BS calculated
from the full (Sanger) sampling; BS value after the slash is BS calculated
from the backbone (plastome) sampling.

For details on methods used to infer the tree, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life.
https://doi.org/10.3389/fpls.2022.909768
}
\keyword{data}
\keyword{trees}
