% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sanger_alignment}
\alias{sanger_alignment}
\title{Aligned (mostly) Sanger DNA sequences used to build Fern Tree of Life (FTOL)}
\format{
Matrix of class "DNAbin"
}
\usage{
sanger_alignment
}
\description{
Concatenation of 7 loci primarily
obtained from Sanger sequencing, including maximal sampling of available
species on GenBank.
}
\details{
Includes 5,611
species and 12,648
bases.

For details on methods used to assemble alignment, see Nitta et al. 2022.
}
\references{
Nitta JH, Schuettpelz E, Ramírez-Barahona S, Iwasaki W. (2022) An
open and continuously updated fern tree of life.
https://doi.org/10.3389/fpls.2022.909768
}
\keyword{alignments}
\keyword{data}
